+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Inverted duration filter in tasks table (#190)
 -   Big refactor to use zustand instead of redux to handle store (#159)
 -   Change files architecture (#159)
+-   Better performances for workflow (#196)
 
 ## [0.40.0] - 2023-03-29